changement de tous les textarea en p

+++ RamosAurore_7_21032021/Backend/controllers/tchat.js
@@ -71,14 +71,16 @@ exports.tchatDelete = (req, res, next) => {
         }
     }).then(tchat => {
         if(tchat.UserId == req.token.userId || req.token.isAdmin){
-            const filename = tchat.attachment.split('/images/')[1];
-            firesystem.unlink(`images/${filename}`, (error => {
-                if(error) 
-                    {console.log(error)}
-                else {
-                    console.log("image effacée");
-                }
-            })) 
+            if(req.file){
+                const filename = tchat.attachment.split('/images/')[1];
+                firesystem.unlink(`images/${filename}`, (error => {
+                    if(error) 
+                        {console.log(error)}
+                    else {
+                        console.log("image effacée");
+                    }
+                })) 
+            }
             Tchat.destroy({ where: { id: req.params.id } })
             .then(() => 
                 res.status(200).json({
